Decouple files for CircleHash64 and CircleHash64f
Closes #4
